requirements.txt
blinker==1.6.2
click==8.1.3
Flask==2.3.1
gpiod==1.5.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
python-dotenv==1.0.0
python-periphery==2.4.1
pytz==2023.3
RPi.bme280==0.2.4
smbus2==0.4.2
tinydb==4.7.1
Werkzeug==2.3.1
zipp==3.15.0
